pyproject.toml
[project]
name = "blog-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "bcrypt>=5.0.0",
    "email-validator>=2.3.0",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "python-slugify>=8.0.4",
    "sqlalchemy>=2.0.48",
    "uvicorn[standard]>=0.41.0",
]
